# expression: implement vectorized evaluation for `builtinArithmeticIntDivideDecimalSig`